# libs/api-client-go/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@runtime.com
    name: Hanzo Industries Inc.
    url: https://www.hanzo.ai
  description: Runtime AI platform API Docs
  title: Runtime
  version: '1.0'
servers:
  - url: http://localhost:3000
paths:
...
  /snapshots:
    get:
      operationId: getAllSnapshots
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Runtime-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: Number of items per page
          explode: true
          in: query
          name: limit
          required: false
          schema:
            type: number
          style: form
        - description: Page number
          explode: true
          in: query
          name: page
          required: false
          schema:
            type: number
          style: form
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedSnapshotsDto'
          description: List of all snapshots with pagination
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: List all snapshots
      tags:
        - snapshots
    post:
      operationId: createSnapshot
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Runtime-Organization-ID
          required: false
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSnapshot'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotDto'
          description: The snapshot has been successfully created.
        '400':
          description: Bad request - Snapshots with tag ":latest" are not allowed
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Create a new snapshot
      tags:
        - snapshots
  /snapshots/{id}:
    delete:
      operationId: removeSnapshot
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Runtime-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: Snapshot ID
          explode: false
          in: path
          name: id
          required: true
          schema:
            type: string
          style: simple
      responses:
        '200':
          description: Snapshot has been deleted
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Delete snapshot
      tags:
        - snapshots
    get:
      operationId: getSnapshot
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Runtime-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: Snapshot ID or name
          explode: false
          in: path
          name: id
          required: true
          schema:
            type: string
          style: simple
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotDto'
          description: The snapshot
        '404':
          description: Snapshot not found
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Get snapshot by ID or name
      tags:
        - snapshots
  /snapshots/{id}/toggle:
    patch:
      operationId: toggleSnapshotState
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Runtime-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: Snapshot ID
          explode: false
          in: path
          name: id
          required: true
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToggleState'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotDto'
          description: Snapshot state has been toggled
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Toggle snapshot state
      tags:
        - snapshots
  /snapshots/{id}/general:
    patch:
      operationId: setSnapshotGeneralStatus
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Runtime-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: Snapshot ID
          explode: false
          in: path
          name: id
          required: true
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetSnapshotGeneralStatusDto'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotDto'
          description: Snapshot general status has been set
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Set snapshot general status
      tags:
        - snapshots
  /snapshots/{id}/build-logs:
    get:
      operationId: getSnapshotBuildLogs
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Runtime-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: Snapshot ID
          explode: false
          in: path
          name: id
          required: true
          schema:
            type: string
          style: simple
        - description: Whether to follow the logs stream
          explode: true
          in: query
          name: follow
          required: false
          schema:
            type: boolean
          style: form
      responses:
        '200':
          description: ''
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Get snapshot build logs
      tags:
        - snapshots
  /snapshots/{id}/activate:
    post:
      operationId: activateSnapshot
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Runtime-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: Snapshot ID
          explode: false
          in: path
          name: id
          required: true
          schema:
            type: string
          style: simple
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotDto'
          description: The snapshot has been successfully activated.
        '400':
          description: "Bad request - Snapshot is already active, not in inactive\
            \ state, or has associated snapshot runners"
        '404':
          description: Snapshot not found
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Activate a snapshot
      tags:
        - snapshots
...
components:
  schemas:
...
    BuildInfo:
      properties:
        dockerfileContent:
          description: The Dockerfile content used for the build
          example: |-
            FROM node:14
            WORKDIR /app
            COPY . .
            RUN npm install
            CMD ["npm", "start"]
          type: string
        contextHashes:
          description: The context hashes used for the build
          example:
            - hash1
            - hash2
          items:
            type: string
          type: array
        createdAt:
          description: The creation timestamp
          format: date-time
          type: string
        updatedAt:
          description: The last update timestamp
          format: date-time
          type: string
      required:
        - createdAt
        - updatedAt
      type: object
...
    CreateBuildInfo:
      properties:
        dockerfileContent:
          description: The Dockerfile content used for the build
          example: |-
            FROM node:14
            WORKDIR /app
            COPY . .
            RUN npm install
            CMD ["npm", "start"]
          type: string
        contextHashes:
          description: The context hashes used for the build
          example:
            - hash1
            - hash2
          items:
            type: string
          type: array
      required:
        - dockerfileContent
      type: object
...
    CreateSnapshot:
      example:
        general: true
        disk: 3
        imageName: ubuntu:22.04
        memory: 1
        buildInfo: ''
        entrypoint: sleep infinity
        name: ubuntu-4vcpu-8ram-100gb
        cpu: 1
        gpu: 0
      properties:
        name:
          description: The name of the snapshot
          example: ubuntu-4vcpu-8ram-100gb
          type: string
        imageName:
          description: The image name of the snapshot
          example: ubuntu:22.04
          type: string
        entrypoint:
          description: The entrypoint command for the snapshot
          example: sleep infinity
          items:
            type: string
          type: array
        general:
          description: Whether the snapshot is general
          type: boolean
        cpu:
          description: CPU cores allocated to the resulting sandbox
          example: 1
          type: integer
        gpu:
          description: GPU units allocated to the resulting sandbox
          example: 0
          type: integer
        memory:
          description: Memory allocated to the resulting sandbox in GB
          example: 1
          type: integer
        disk:
          description: Disk space allocated to the sandbox in GB
          example: 3
          type: integer
        buildInfo:
          allOf:
            - $ref: '#/components/schemas/CreateBuildInfo'
          description: Build information for the snapshot
      required:
        - name
      type: object
    SnapshotState:
      enum:
        - build_pending
        - building
        - pending
        - pulling
        - pending_validation
        - validating
        - active
        - inactive
        - error
        - build_failed
        - removing
      type: string
    SnapshotDto:
      example:
        imageName: imageName
        buildInfo: ''
        cpu: 6.027456183070403
        gpu: 1.4658129805029452
        enabled: true
        organizationId: organizationId
        general: true
        createdAt: 2000-01-23T04:56:07.000+00:00
        disk: 5.637376656633329
        size: 0.8008281904610115
        mem: 5.962133916683182
        lastUsedAt: 2000-01-23T04:56:07.000+00:00
        entrypoint:
          - entrypoint
          - entrypoint
        errorReason: errorReason
        name: name
        id: id
        state: ''
        updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          type: string
        organizationId:
          type: string
        general:
          type: boolean
        name:
          type: string
        imageName:
          type: string
        enabled:
          type: boolean
        state:
          allOf:
            - $ref: '#/components/schemas/SnapshotState'
        size:
          nullable: true
          type: number
        entrypoint:
          items:
            type: string
          nullable: true
          type: array
        cpu:
          type: number
        gpu:
          type: number
        mem:
          type: number
        disk:
          type: number
        errorReason:
          nullable: true
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        lastUsedAt:
          format: date-time
          nullable: true
          type: string
        buildInfo:
          allOf:
            - $ref: '#/components/schemas/BuildInfo'
          description: Build information for the snapshot
      required:
        - cpu
        - createdAt
        - disk
        - enabled
        - entrypoint
        - errorReason
        - general
        - gpu
        - id
        - lastUsedAt
        - mem
        - name
        - size
        - state
        - updatedAt
      type: object
    ToggleState:
      example:
        enabled: true
      properties:
        enabled:
          description: Enable or disable the snapshot/tag
          example: true
          type: boolean
      required:
        - enabled
      type: object
    PaginatedSnapshotsDto:
      example:
        total: 2.3021358869347655
        totalPages: 9.301444243932576
        page: 7.061401241503109
        items:
          - imageName: imageName
            buildInfo: ''
            cpu: 6.027456183070403
            gpu: 1.4658129805029452
            enabled: true
            organizationId: organizationId
            general: true
            createdAt: 2000-01-23T04:56:07.000+00:00
            disk: 5.637376656633329
            size: 0.8008281904610115
            mem: 5.962133916683182
            lastUsedAt: 2000-01-23T04:56:07.000+00:00
            entrypoint:
              - entrypoint
              - entrypoint
            errorReason: errorReason
            name: name
            id: id
            state: ''
            updatedAt: 2000-01-23T04:56:07.000+00:00
          - imageName: imageName
            buildInfo: ''
            cpu: 6.027456183070403
            gpu: 1.4658129805029452
            enabled: true
            organizationId: organizationId
            general: true
            createdAt: 2000-01-23T04:56:07.000+00:00
            disk: 5.637376656633329
            size: 0.8008281904610115
            mem: 5.962133916683182
            lastUsedAt: 2000-01-23T04:56:07.000+00:00
            entrypoint:
              - entrypoint
              - entrypoint
            errorReason: errorReason
            name: name
            id: id
            state: ''
            updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        items:
          items:
            $ref: '#/components/schemas/SnapshotDto'
          type: array
        total:
          type: number
        page:
          type: number
        totalPages:
          type: number
      required:
        - items
        - page
        - total
        - totalPages
      type: object
    SetSnapshotGeneralStatusDto:
      example:
        general: true
      properties:
        general:
          description: Whether the snapshot is general
          example: true
          type: boolean
      required:
        - general
      type: object
...
  securitySchemes:
    bearer:
      bearerFormat: JWT
      description: API Key access
      scheme: bearer
      type: http
    oauth2:
      openIdConnectUrl: http://localhost:3000/.well-known/openid-configuration
      type: openIdConnect